[#81] Support for overwriting clangd options
- Get rid of ClangdOptionsDefaultsImpl
