+++ bundles/org.eclipse.cdt.lsp.clangd/OSGI-INF/org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaults.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaults">
+   <service>
+      <provide interface="org.eclipse.cdt.lsp.clangd.ClangdOptions"/>
+   </service>
+   <implementation class="org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaults"/>
+   <property name="service.ranking" type="Integer" value="0"/>
+   <property name="isDefaultOptions" type="Boolean" value="true"/>
+</scr:component>
