--- bundles/org.eclipse.cdt.lsp.clangd/OSGI-INF/org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaultsImpl.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaultsImpl">
-   <service>
-      <provide interface="org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaults"/>
-   </service>
-   <implementation class="org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaultsImpl"/>
-   <property name="service.ranking" type="Integer" value="0"/>
-</scr:component>
