--- examples/org.eclipse.cdt.lsp.examples.preferences/OSGI-INF/org.eclipse.cdt.lsp.examples.preferences.MyClangdOptionsDefaultsImpl.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.3.0" name="org.eclipse.cdt.lsp.examples.preferences.MyClangdOptionsDefaultsImpl">
-   <service>
-      <provide interface="org.eclipse.cdt.lsp.clangd.ClangdOptionsDefaults"/>
-   </service>
-   <property name="service.ranking" type="Integer" value="100"/>
-   <implementation class="org.eclipse.cdt.lsp.examples.preferences.MyClangdOptionsDefaultsImpl"/>
-</scr:component>
